Refactor to facilitate per-component entries
This is a pretty big refactor to address #2 without the need for a
configuration file. This approach, however, assumes that we will have to
implement per-programming-language support for `unclog`. Nothing stops
us, however, from, in the future, building support for a generic
configuration file for languages we don't understand.

This commit includes:
1. Code updates
2. Integration test update

Signed-off-by: Thane Thomson <connect@thanethomson.com>